+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
   - 🌟 [@mixte/use] 方法 `useRequest` 新增 `successCount` 和 `clearSuccessCount` 返回值, 统计请求成功次数
+  - 💄 [@mixte/use] 优化 `useRequestReactive` 方法类型定义
 
 ## [v1.6.3]
   - 📅 2023-11-22